pyproject.toml
[tool.poetry]
name = "gcore-api"
version = "1.0.0"
description = "Command-line tool for Gcore API integration"
authors = ["DevopsX <devopsx@example.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/mystery000/gcore-api"
repository = "https://github.com/mystery000/gcore-api"
documentation = "https://github.com/mystery000/gcore-api#readme"
keywords = ["gcore", "cdn", "dns", "storage", "loadbalancer", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities"
]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.8"
requests = "^2.32.3"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
black = "^24.1.1"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.8.0"
types-requests = "^2.31.0.20240125"
types-PyYAML = "^6.0.12.12"

[tool.poetry.scripts]
gcore = "gcore_api.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py39"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=gcore_api --cov-report=term-missing"
